feat: create the knownChainsAtom
