+++ apps/namadillo/src/App/Ibc/IbcTransfer.tsx
@@ -7,18 +7,6 @@ import { wallets } from "integrations";
 import { useEffect, useMemo, useState } from "react";
 import { ChainRegistryEntry, WalletProvider } from "types";
 
-import * as celestia from "chain-registry/mainnet/celestia";
-import * as cosmos from "chain-registry/mainnet/cosmoshub";
-import * as dydx from "chain-registry/mainnet/dydx";
-import * as osmosis from "chain-registry/mainnet/osmosis";
-import * as stargaze from "chain-registry/mainnet/stargaze";
-
-import * as celestiaTestnet from "chain-registry/testnet/celestiatestnet3";
-import * as cosmosTestnet from "chain-registry/testnet/cosmoshubtestnet";
-import * as dydxTestnet from "chain-registry/testnet/dydxtestnet";
-import * as osmosisTestnet from "chain-registry/testnet/osmosistestnet4";
-import * as stargazeTestnet from "chain-registry/testnet/stargazetestnet";
-
 import { allDefaultAccountsAtom } from "atoms/accounts";
 
 import { useAtom, useAtomValue } from "jotai";
@@ -28,9 +16,9 @@ import { Asset, Chain } from "@chain-registry/types";
 import {
   assetBalanceAtomFamily,
   ibcTransferAtom,
+  knownChainsAtom,
   selectedIBCChainAtom,
 } from "atoms/integrations";
-import { settingsAtom } from "atoms/settings";
 import { basicConvertToKeplrChain } from "utils/integration";
 
 const keplr = (window as KeplrWindow).keplr!;
@@ -38,24 +26,8 @@ const keplr = (window as KeplrWindow).keplr!;
 //TODO: we need to find a good way to manage IBC channels
 const namadaChannelId = "channel-4353";
 
-const mainnetChains: Record<string, ChainRegistryEntry> = {
-  [celestia.chain.chain_id]: celestia,
-  [cosmos.chain.chain_id]: cosmos,
-  [dydx.chain.chain_id]: dydx,
-  [osmosis.chain.chain_id]: osmosis,
-  [stargaze.chain.chain_id]: stargaze,
-};
-
-const testnetChains: Record<string, ChainRegistryEntry> = {
-  [cosmosTestnet.chain.chain_id]: cosmosTestnet,
-  [celestiaTestnet.chain.chain_id]: celestiaTestnet,
-  [dydxTestnet.chain.chain_id]: dydxTestnet,
-  [osmosisTestnet.chain.chain_id]: osmosisTestnet,
-  [stargazeTestnet.chain.chain_id]: stargazeTestnet,
-};
-
 export const IbcTransfer: React.FC = () => {
-  const settings = useAtomValue(settingsAtom);
+  const knownChains = useAtomValue(knownChainsAtom);
   const [chainId, setChainId] = useAtom(selectedIBCChainAtom);
   const [registry, setRegistry] = useState<ChainRegistryEntry>();
   const [sourceAddress, setSourceAddress] = useState<string | undefined>();
@@ -76,12 +48,6 @@ export const IbcTransfer: React.FC = () => {
       sourceAddress,
     })
   );
-
-  const knownChains = useMemo(() => {
-    return settings.enableTestnets ?
-        { ...mainnetChains, ...testnetChains }
-      : mainnetChains;
-  }, [settings.enableTestnets]);
 
   const namadaAddress = useMemo(() => {
     return (
